More HCI improvements

See merge request cse1105/2024-2025/teams/csep-team-67!50

+++ client/src/main/java/client/controllers/CollectionCtrl.java
@@ -1,36 +1,26 @@
 package client.controllers;
 
-import client.MyFXML;
-import client.MyModule;
 import client.scenes.DashboardCtrl;
 import client.scenes.EditCollectionsCtrl;
 import client.ui.DialogStyler;
 import client.utils.Config;
 import client.utils.ServerUtils;
 import com.google.inject.Inject;
-import com.google.inject.Injector;
 import commons.Collection;
 import commons.Note;
 import jakarta.ws.rs.ClientErrorException;
 import javafx.beans.binding.Bindings;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import javafx.scene.Scene;
 import javafx.scene.control.*;
-import javafx.stage.Modality;
-import javafx.stage.Stage;
-import javafx.stage.StageStyle;
+import lombok.Getter;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static com.google.inject.Guice.createInjector;
 
 
 public class CollectionCtrl {
-
-    private static final Injector INJECTOR = createInjector(new MyModule());
-    private static final MyFXML FXML = new MyFXML(INJECTOR);
 
     // Utilities
     private final ServerUtils server;
@@ -44,7 +34,7 @@ public class CollectionCtrl {
     private ListView collectionView;
     private TreeView treeView;
     private MenuButton currentCollectionTitle;
-    private ToggleGroup collectionSelect;
+    @Getter private ToggleGroup collectionSelect;
     private MenuItem allNotesButton;
     private MenuItem editCollectionTitle;
     private Button deleteCollectionButton;
@@ -65,7 +55,6 @@ public class CollectionCtrl {
     public void setReferences(ListView collectionView,
                               TreeView treeView,
                               MenuButton currentCollectionTitle,
-                              ToggleGroup collectionSelect,
                               MenuItem allNotesButton,
                               MenuItem editCollectionTitle,
                               Button deleteCollectionButton,
@@ -312,12 +301,8 @@ public class CollectionCtrl {
             currentCollectionTitle.setText("All Notes");
             collectionView.setVisible(false);
             treeView.setVisible(true);
-            dashboardCtrl.refreshTreeView();
             collectionView.getSelectionModel().clearSelection();
         } else {
-            server.getWebSocketURL(currentCollection.serverURL);
-            dashboardCtrl.noteAdditionSync();
-            dashboardCtrl.noteDeletionSync();
             collectionNotes = FXCollections.observableArrayList(
                     allNotes.stream()
                             .filter(note -> note.collection.equals(currentCollection))
@@ -339,37 +324,42 @@ public class CollectionCtrl {
         return collectionNotes!=null?collectionNotes : FXCollections.observableArrayList();
     }
 
-
-    public Collection deleteCollection(Collection currentCollection,
-                                       List<Collection> collections,
-                                       ObservableList<Note> collectionNotes,
-                                       ObservableList<Note> allNotes)
-    {
-        if (!showDeleteConfirmation()) return currentCollection;
-
-        List<Note> notesToDelete = collectionNotes.stream().toList();
+    public void removeCollectionFromClient(boolean delete, Collection collection, List<Collection> collections, ObservableList<Note> collectionNotes, ObservableList<Note> allNotes) {
+        List<Note> notesToDelete = allNotes.stream()
+                .filter(note -> note.collection.equals(collection))
+                .collect(Collectors.toList());
         for (Note n : notesToDelete) {
-            noteCtrl.deleteNote(n,collectionNotes,allNotes);
+            if (delete) noteCtrl.deleteNote(n,collectionNotes,allNotes);
         }
+
+        Collection previousCollection = collections.stream().filter(c -> !c.equals(collection)).findFirst().orElse(null);
+        dashboardCtrl.setDefaultCollection(previousCollection);
+        config.setDefaultCollection(previousCollection);
+
         // delete collection from server
-        server.deleteCollection(currentCollection);
+        if (delete) server.deleteCollection(collection);
         // delete collection from config file
-        collections.remove(currentCollection);
+        collections.remove(collection);
 
         config.writeAllToFile(collections);
-        // delete collection from collections menu
-        //currentCollectionTitle.getItems().remove(
-//.getSelectedToggle()
-        //);
-        return null;
     }
 
-    private boolean showDeleteConfirmation() {
+    public boolean showDeleteConfirmation() {
         return dialogStyler.createStyledAlert(
                 Alert.AlertType.CONFIRMATION,
                 "Delete collection",
                 "Delete collection",
                 "Are you sure you want to delete this collection? All notes in the collection will be deleted as well."
+        ).showAndWait().filter(b -> b == ButtonType.OK).isPresent();
+    }
+
+    public boolean showForgetConfirmation() {
+        return dialogStyler.createStyledAlert(
+                Alert.AlertType.CONFIRMATION,
+                "Forget collection",
+                "Forget collection",
+                "Are you sure you want to forget this collection?" +
+                        "\nYou will lose access to it's notes, but may reconnect to it later."
         ).showAndWait().filter(b -> b == ButtonType.OK).isPresent();
     }
 
@@ -424,52 +414,16 @@ public class CollectionCtrl {
         RadioMenuItem radioMenuItem = dashboardCtrl.createCollectionButton(addedCollection, currentCollectionTitle, collectionSelect);
         collectionSelect.selectToggle(radioMenuItem);
 
-        return addedCollection;
+        return currentCollection == null ? currentCollection : addedCollection;
     }
 
     public void addCollection(){
-        var editCollections = FXML.load(EditCollectionsCtrl.class, "client", "scenes", "EditCollections.fxml");
-
-        Stage popupStage = new Stage();
-        popupStage.initModality(Modality.APPLICATION_MODAL); // Block interaction with main window
-        popupStage.initStyle(StageStyle.TRANSPARENT); // Make window transparent
-        popupStage.setTitle("Popup Window");
-
-        Scene scene = new Scene(editCollections.getValue());
-        scene.setFill(javafx.scene.paint.Color.TRANSPARENT); // Transparent scene
-
-        popupStage.setScene(scene);
-        popupStage.getScene().getStylesheets().add(getClass().getResource("/css/styles.css").toExternalForm());
-
-        // Set the stage in controller for dragging
-        EditCollectionsCtrl controller = (EditCollectionsCtrl) editCollections.getKey();
-        controller.setReferences(popupStage, this, dashboardCtrl, server, noteCtrl, config, dialogStyler);
-        controller.setCollectionList(dashboardCtrl.getCollections());
-        controller.addCollection();
-
-        popupStage.showAndWait();
+        EditCollectionsCtrl editCollectionsCtrl = dashboardCtrl.getMainCtrl().showEditCollections();
+        editCollectionsCtrl.addCollection();
     }
 
     public void editCollections() {
-        var editCollections = FXML.load(EditCollectionsCtrl.class, "client", "scenes", "EditCollections.fxml");
-
-        Stage popupStage = new Stage();
-        popupStage.initModality(Modality.APPLICATION_MODAL); // Block interaction with main window
-        popupStage.initStyle(StageStyle.TRANSPARENT); // Make window transparent
-        popupStage.setTitle("Popup Window");
-
-        Scene scene = new Scene(editCollections.getValue());
-        scene.setFill(javafx.scene.paint.Color.TRANSPARENT); // Transparent scene
-
-        popupStage.setScene(scene);
-        popupStage.getScene().getStylesheets().add(getClass().getResource("/css/styles.css").toExternalForm());
-
-        // Set the stage in controller for dragging
-        EditCollectionsCtrl controller = (EditCollectionsCtrl) editCollections.getKey();
-        controller.setReferences(popupStage, this, dashboardCtrl, server, noteCtrl, config, dialogStyler);
-        controller.setCollectionList(dashboardCtrl.getCollections());
-
-        popupStage.showAndWait();
+        dashboardCtrl.getMainCtrl().showEditCollections();
     }
